pyproject.toml
[project]
name = "vgi-rpc"
version = "0.1.0"
description = "Vector Gateway Interface - RPC framework based on Apache Arrow"
readme = "README.md"
requires-python = ">=3.13"
license = {file = "LICENSE.md"}
authors = [{ name = "Query.farm" }]
keywords = ["rpc", "arrow", "ipc", "serialization", "streaming"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = ["pyarrow>=14.0", "tzdata>=2024.1; sys_platform == 'win32'"]

[project.urls]
Homepage = "https://github.com/Query-farm/vgi-rpc-python"
Repository = "https://github.com/Query-farm/vgi-rpc-python"
Issues = "https://github.com/Query-farm/vgi-rpc-python/issues"
Documentation = "https://query-farm.github.io/vgi-rpc-python/"

[project.optional-dependencies]
http = ["falcon>=3.0", "httpx>=0.24", "waitress>=2.0"]
s3 = ["boto3>=1.28"]
gcs = ["google-cloud-storage>=2.10"]
cli = ["typer>=0.9", "httpx>=0.24"]
external = ["aiohttp>=3.9", "tenacity>=8.0", "zstandard>=0.20"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]

[project.scripts]
vgi-rpc = "vgi_rpc.cli:app"
vgi-rpc-conformance = "vgi_rpc.conformance._cli:main"
vgi-rpc-test = "vgi_rpc.conformance._test_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "D", "RUF", "PERF"]
ignore = [
  "D203",   # incompatible with D211 (no-blank-line-before-class)
  "D213",   # incompatible with D212 (multi-line-summary-first-line)
  "RUF022", # __all__ is intentionally grouped by category, not sorted
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "falcon.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "waitress.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "boto3.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "moto.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aioresponses.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "typer.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest_benchmark.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
# Protocol classes are passed as type markers (not instantiated) to serve_pipe/connect/http_connect;
# type[P] correctly triggers type-abstract for abstract Protocol classes.
disable_error_code = ["type-abstract"]

[tool.ty.rules]
unused-type-ignore-comment = "ignore"  # these comments are required by mypy

[tool.ty.environment]
python-version = "3.13"

[tool.pytest.ini_options]
addopts = "--mypy --ruff --timeout=50 --benchmark-disable -m 'not benchmark'"
markers = ["benchmark: performance benchmarks (deselected by default)"]
testpaths = ["tests"]

[tool.coverage.run]
# Each subprocess writes to its own data file (.coverage.<hostname>.<pid>)
parallel = true
# Handle SIGTERM gracefully (writes coverage data before exit)
sigterm = true
# Measure branch coverage
branch = true
# What to measure
source = ["vgi_rpc"]
# Omit test files and examples from coverage
omit = ["vgi_rpc/examples/*", "tests/*"]

[tool.coverage.report]
# Fail if coverage drops below this threshold
fail_under = 80
# Show missing lines in the report
show_missing = true
# Exclude common patterns from coverage
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
  "@overload",
  "raise NotImplementedError",
  "\\.\\.\\.",                  # Ellipsis in stub files
]

[tool.coverage.html]
directory = "htmlcov"

[tool.mutmut]
paths_to_mutate = ["vgi_rpc/utils.py"]
pytest_add_cli_args_test_selection = ["tests/"]
also_copy = ["vgi_rpc/"]
pytest_add_cli_args = ["--timeout=50", "--no-header", "-o", "addopts=", "-o", "pythonpath=.", "--ignore=tests/test_external_fetch.py"]

[dependency-groups]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
    "mkdocs-section-index>=0.3",
    "mkdocs-d2-plugin>=1.4",
]
dev = [
    "aioresponses>=0.7",
    "boto3>=1.28",
    "falcon>=3.0",
    "google-cloud-storage>=2.10",
    "httpx>=0.24",
    "lxml>=6.0.2",
    "moto[s3]>=4.0",
    "mutmut>=3.4.0",
    "mypy>=1.0",
    "pyarrow-stubs>=17.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-examples>=0.0.10",
    "pytest-mypy>=0.10",
    "pytest-ruff>=0.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.1",
    "ty>=0.0.8",
    "typer>=0.9",
    "waitress>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "pytest-benchmark>=4.0",
]
